feat(platform): migrate to ESP-IDF 5.1 (pioarduino)
Phase 0 of ESP-IDF migration - platform upgrade with minimal fixes:

- Switch from espressif32@6.9.0 to pioarduino (ESP-IDF 5.5.2, Arduino 3.3.7)
- Update WDT API to ESP-IDF 5.x (esp_task_wdt_config_t + reconfigure)
- Fix format specifiers for uint32_t (%lu -> %u with cast)
- Fix include paths for arduino_compat.h
- Fix lib_extra_dirs for Network library discovery
- Add [[maybe_unused]] for unused TAG variables

Tested on T-Deck Plus: boot, LVGL, LoRa RX, GPS, map (raster+NAV) all working.

+++ src/lora_utils.cpp
@@ -27,7 +27,7 @@
 #include "display.h"
 #include "station_utils.h"
 #include "storage_utils.h"
-#include "compat/arduino_compat.h"
+#include "../compat/arduino_compat.h"
 #ifdef USE_LVGL_UI
 #include "lvgl_ui.h"
 #endif
@@ -230,18 +230,18 @@ namespace LoRa_Utils {
             // Board-specific frequency validation
             if (currentLoRaType->frequency < LORA_FREQ_MIN || currentLoRaType->frequency > LORA_FREQ_MAX) {
                 ESP_LOGE(TAG, "Frequency %ld Hz out of range (%ld-%ld)",
-                    currentLoRaType->frequency, (long)LORA_FREQ_MIN, (long)LORA_FREQ_MAX);
+                    (long)currentLoRaType->frequency, (long)LORA_FREQ_MIN, (long)LORA_FREQ_MAX);
                 return;
             }
         #else
             // Generic frequency validation
             if (currentLoRaType->frequency < 100000000 || currentLoRaType->frequency > 1000000000) {
-                ESP_LOGE(TAG, "Invalid frequency value: %ld", currentLoRaType->frequency);
+                ESP_LOGE(TAG, "Invalid frequency value: %ld", (long)currentLoRaType->frequency);
                 return;
             }
         #endif
         if (currentLoRaType->signalBandwidth < 1000 || currentLoRaType->signalBandwidth > 1000000) {
-            ESP_LOGE(TAG, "Invalid bandwidth value: %ld", currentLoRaType->signalBandwidth);
+            ESP_LOGE(TAG, "Invalid bandwidth value: %ld", (long)currentLoRaType->signalBandwidth);
             return;
         }
